[Kinect] Fix acquiring of RGB frames on the client
Summary:
RGB frames are now constructed correctly.
However, fixing depth and undistorted RGB frames is still WIP.

`displayWhatYouSee` is still being used for debugging. Yes, it's clownly.
However, getting the threads to dispay the frames and not crash and burn on the slave is just too painful.
Will remove once depth and undistorted is fixed.

Test Plan:
Plug in the Kinect
Comment out call to displayGreyCodes on the master
Uncomment the call to displayWhatYouSee on the master
`./build/master/master`
`./build/slave/slave`
Observe that the RGB image captured by the Kinect gets displated on the slave

Reviewers: piotr, hanq, michsien, nand

Reviewed By: nand

Differential Revision: https://ec2-52-30-7-90.eu-west-1.compute.amazonaws.com/D59