Added custom Kinect logger
Test Plan:
Run the slave with kinect enabled. Play around with log-level and log-filename flags to see different amounts of log msgs being logged and in different places (file/console).
Run with kinect disabled to see that it pretends that it is logging and doesn't break.
(Should be also tested with Kinect connected to the machine to confirm it doesn't break then.)

Reviewers: ilijar, hanq, piotr, nand

Reviewed By: nand

Differential Revision: https://ec2-52-30-7-90.eu-west-1.compute.amazonaws.com/D74