Fixed broken gray code generation
Summary:
Corrected the way we generate and store gray code images to allow for
correct limiting of gray code levels we display.

Test Plan:
Run master and slave and see that now gra code levels are actually
limited at the end (not at the front) i.e. we display levels 0-5, not
e.g. 4-9.

Reviewers: ilijar, nand, piotr, hanq

Reviewed By: hanq

Differential Revision: https://ec2-52-30-7-90.eu-west-1.compute.amazonaws.com/D97